requirements.txt
aiobotocore==2.21.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aioitertools==0.12.0
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
apache-beam==2.63.0
argcomplete==3.6.1
arraylake==0.16.0
asttokens==3.0.0
attrs==25.3.0
black==25.1.0
bleach==6.2.0
blessed==1.21.0
bokeh==3.7.0
boto==2.49.0
boto3==1.37.1
botocore==1.37.1
Bottleneck==1.4.2
cachetools==5.5.2
cachey==0.2.1
Cartopy==0.24.1
cdsapi==0.7.5
certifi==2025.1.31
cffi==1.17.1
cftime==1.6.4.post1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorcet==3.1.0
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.1
crc32c==2.7.1
crcmod==1.7
cryptography==43.0.3
cycler==0.12.1
dask==2025.3.0
datapi==0.3.0
datashader==0.17.0
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
dill==0.3.1.1
distributed==2025.3.0
dnspython==2.7.0
docopt==0.6.2
donfig==0.8.1.post1
# Editable install with no version control (early-stopping-pytorch==1.0.10)
-e /homes/nb324/Documents/masters_project_encoder_sparse_data_robust_imp_dec/early-stopping-pytorch
email_validator==2.2.0
executing==2.2.0
fastavro==1.10.0
fasteners==0.19
filelock==3.18.0
flake8==7.2.0
flox==0.10.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
gcs-oauth2-boto-plugin==3.2
gcsfs==2025.3.0
google-api-core==2.24.2
google-apitools==0.5.32
google-auth==2.17.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.0
google-reauth==0.1.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
gpustat==1.1.1
gpytorch==1.14
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.65.5
gsutil==5.33
h11==0.14.0
h5netcdf==1.6.1
h5py==3.13.0
hdfs==2.7.3
HeapDict==1.0.1
holoviews==1.20.2
httpcore==1.0.7
httplib2==0.20.4
httpx==0.27.2
humanize==4.9.0
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
isort==6.0.1
jaxtyping==0.3.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jsonpickle==3.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
lightning-utilities==0.14.2
linear-operator==0.6
linkify-it-py==2.0.3
llvmlite==0.44.0
locket==1.0.0
lz4==4.4.3
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
multipledispatch==1.0.0
multiurl==0.3.5
mypy_extensions==1.1.0
narwhals==1.32.0
nc-time-axis==1.4.1
nest-asyncio==1.6.0
netCDF4==1.7.2
networkx==3.4.2
numba==0.61.0
numbagg==0.9.0
numcodecs==0.15.1
numpy==2.1.3
numpy-groupies==0.11.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauth2client==4.1.3
oauthlib==3.2.2
objsize==0.7.1
opt_einsum==3.4.0
optuna==4.2.1
orjson==3.10.15
packaging==24.2
pandas==2.2.3
panel==1.6.2
param==2.2.0
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
pooch==1.8.2
prefetch_generator==1.0.3
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.13.0
pycparser==2.22
pyct==0.5.0
pydantic==2.9.2
pydantic_core==2.23.4
pydot==1.4.2
pyflakes==3.3.2
Pygments==2.19.1
pymongo==4.11.3
pynndescent==0.5.13
pyOpenSSL==24.2.1
pyparsing==3.2.2
pyproj==3.7.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
pytorch-metric-learning==2.8.1
pytz==2025.1
pyu2f==0.1.5
pyviz_comms==3.0.4
PyYAML==6.0.2
pyzmq==26.3.0
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
retry_decorator==1.1.1
rich==13.9.4
rpds-py==0.23.1
rsa==4.7.2
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3fs==2025.3.0
s3transfer==0.11.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==78.0.1
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sparse==0.15.5
SQLAlchemy==2.0.40
sqlitedict==2.1.0
stack-data==0.6.3
structlog==24.4.0
sympy==1.13.1
tblib==3.0.0
tenacity==8.5.0
threadpoolctl==3.6.0
tifffile==2025.3.30
tokenize_rt==6.2.0
toolz==1.0.0
torch==2.6.0
torch-geometric==2.6.1
torch-summary==1.4.5
torchda @ git+https://github.com/acse-jm122/torchda.git@1b02e2990796d790a5cf4b46109d273aeb9ff892
torchdiffeq==0.2.5
torchmetrics==1.7.0
torchvision==0.21.0
torchviz==0.0.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.2.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.2
uc-micro-py==1.0.3
umap-learn==0.5.7
urllib3==2.3.0
wadler_lindig==0.1.5
wcwidth==0.2.13
webencodings==0.5.1
wrapt==1.17.2
xarray==2025.3.0
xbatcher==0.4.0
xyzservices==2025.1.0
yarl==1.18.3
zarr==3.0.6
zict==3.0.0
zstandard==0.23.0
